+++ packages/firestore/src/api/database.ts
@@ -1000,43 +1000,91 @@ export class DocumentReference implements firestore.DocumentReference {
     };
   }
 
-  get(): Promise<firestore.DocumentSnapshot> {
-    validateExactNumberOfArgs('DocumentReference.get', arguments, 0);
+  get(options?: firestore.GetOptions): Promise<firestore.DocumentSnapshot> {
+    validateOptionNames('DocumentReference.get', options, ['source']);
+    if (options) {
+      validateNamedOptionalPropertyEquals(
+        'DocumentReference.get',
+        'options',
+        'source',
+        options.source,
+        ['default', 'server', 'cache']
+      );
+    }
     return new Promise(
       (resolve: Resolver<firestore.DocumentSnapshot>, reject: Rejecter) => {
-        const unlisten = this.onSnapshotInternal(
-          {
-            includeQueryMetadataChanges: true,
-            includeDocumentMetadataChanges: true,
-            waitForSyncWhenOnline: true
-          },
-          {
-            next: (snap: firestore.DocumentSnapshot) => {
-              // Remove query first before passing event to user to avoid
-              // user actions affecting the now stale query.
-              unlisten();
+        if (options && options.source === 'cache') {
+          this.firestore
+            .ensureClientConfigured()
+            .getDocumentFromLocalCache(this._key)
+            .then((doc: Document) => {
+              resolve(
+                new DocumentSnapshot(
+                  this.firestore,
+                  this._key,
+                  doc,
+                  /*fromCache=*/ true
+                )
+              );
+            }, reject);
+        } else {
+          this.getViaSnapshotListener(resolve, reject, options);
+        }
+      }
+    );
+  }
 
-              if (!snap.exists && snap.metadata.fromCache) {
-                // TODO(dimond): If we're online and the document doesn't
-                // exist then we resolve with a doc.exists set to false. If
-                // we're offline however, we reject the Promise in this
-                // case. Two options: 1) Cache the negative response from
-                // the server so we can deliver that even when you're
-                // offline 2) Actually reject the Promise in the online case
-                // if the document doesn't exist.
-                reject(
-                  new FirestoreError(
-                    Code.ABORTED,
-                    'Failed to get document because the client is ' + 'offline.'
-                  )
-                );
-              } else {
-                resolve(snap);
-              }
-            },
-            error: reject
+  private getViaSnapshotListener(
+    resolve: Resolver<firestore.DocumentSnapshot>,
+    reject: Rejecter,
+    options?: firestore.GetOptions
+  ): void {
+    const unlisten = this.onSnapshotInternal(
+      {
+        includeQueryMetadataChanges: true,
+        includeDocumentMetadataChanges: true,
+        waitForSyncWhenOnline: true
+      },
+      {
+        next: (snap: firestore.DocumentSnapshot) => {
+          // Remove query first before passing event to user to avoid
+          // user actions affecting the now stale query.
+          unlisten();
+
+          if (!snap.exists && snap.metadata.fromCache) {
+            // TODO(dimond): If we're online and the document doesn't
+            // exist then we resolve with a doc.exists set to false. If
+            // we're offline however, we reject the Promise in this
+            // case. Two options: 1) Cache the negative response from
+            // the server so we can deliver that even when you're
+            // offline 2) Actually reject the Promise in the online case
+            // if the document doesn't exist.
+            reject(
+              new FirestoreError(
+                Code.UNAVAILABLE,
+                'Failed to get document because the client is ' + 'offline.'
+              )
+            );
+          } else if (
+            snap.exists &&
+            snap.metadata.fromCache &&
+            options &&
+            options.source === 'server'
+          ) {
+            reject(
+              new FirestoreError(
+                Code.UNAVAILABLE,
+                'Failed to get document from server. (However, this ' +
+                  'document does exist in the local cache. Run again ' +
+                  'without setting source to "server" to ' +
+                  'retrieve the cached document.)'
+              )
+            );
+          } else {
+            resolve(snap);
           }
-        );
+        },
+        error: reject
       }
     );
   }
@@ -1619,27 +1667,60 @@ export class Query implements firestore.Query {
     };
   }
 
-  get(): Promise<firestore.QuerySnapshot> {
-    validateExactNumberOfArgs('Query.get', arguments, 0);
+  get(options?: firestore.GetOptions): Promise<firestore.QuerySnapshot> {
+    validateBetweenNumberOfArgs('Query.get', arguments, 0, 1);
     return new Promise(
       (resolve: Resolver<firestore.QuerySnapshot>, reject: Rejecter) => {
-        const unlisten = this.onSnapshotInternal(
-          {
-            includeDocumentMetadataChanges: false,
-            includeQueryMetadataChanges: true,
-            waitForSyncWhenOnline: true
-          },
-          {
-            next: (result: firestore.QuerySnapshot) => {
-              // Remove query first before passing event to user to avoid
-              // user actions affecting the now stale query.
-              unlisten();
+        if (options && options.source === 'cache') {
+          this.firestore
+            .ensureClientConfigured()
+            .getDocumentsFromLocalCache(this._query)
+            .then((viewSnap: ViewSnapshot) => {
+              resolve(new QuerySnapshot(this.firestore, this._query, viewSnap));
+            }, reject);
+        } else {
+          this.getViaSnapshotListener(resolve, reject, options);
+        }
+      }
+    );
+  }
 
-              resolve(result);
-            },
-            error: reject
+  private getViaSnapshotListener(
+    resolve: Resolver<firestore.QuerySnapshot>,
+    reject: Rejecter,
+    options?: firestore.GetOptions
+  ): void {
+    const unlisten = this.onSnapshotInternal(
+      {
+        includeDocumentMetadataChanges: false,
+        includeQueryMetadataChanges: true,
+        waitForSyncWhenOnline: true
+      },
+      {
+        next: (result: firestore.QuerySnapshot) => {
+          // Remove query first before passing event to user to avoid
+          // user actions affecting the now stale query.
+          unlisten();
+
+          if (
+            result.metadata.fromCache &&
+            options &&
+            options.source === 'server'
+          ) {
+            reject(
+              new FirestoreError(
+                Code.UNAVAILABLE,
+                'Failed to get documents from server. (However, these ' +
+                  'documents may exist in the local cache. Run again ' +
+                  'without setting source to "server" to ' +
+                  'retrieve the cached documents.)'
+              )
+            );
+          } else {
+            resolve(result);
           }
-        );
+        },
+        error: reject
       }
     );
   }