+++ packages/firestore/src/protos/firestore_proto_api.d.ts
@@ -33,7 +33,8 @@ export declare type FieldFilterOp =
   | 'LESS_THAN_OR_EQUAL'
   | 'GREATER_THAN'
   | 'GREATER_THAN_OR_EQUAL'
-  | 'EQUAL';
+  | 'EQUAL'
+  | 'ARRAY_CONTAINS';
 export interface IFieldFilterOpEnum {
   OPERATOR_UNSPECIFIED: FieldFilterOp;
   LESS_THAN: FieldFilterOp;
@@ -41,6 +42,7 @@ export interface IFieldFilterOpEnum {
   GREATER_THAN: FieldFilterOp;
   GREATER_THAN_OR_EQUAL: FieldFilterOp;
   EQUAL: FieldFilterOp;
+  ARRAY_CONTAINS: FieldFilterOp;
   values(): Array<FieldFilterOp>;
 }
 export declare const FieldFilterOpEnum: IFieldFilterOpEnum;
@@ -64,15 +66,6 @@ export interface IIndexFieldModeEnum {
   values(): Array<IndexFieldMode>;
 }
 export declare const IndexFieldModeEnum: IIndexFieldModeEnum;
-export declare type IndexOperationMetadataOperationType =
-  | 'OPERATION_TYPE_UNSPECIFIED'
-  | 'CREATING_INDEX';
-export interface IIndexOperationMetadataOperationTypeEnum {
-  OPERATION_TYPE_UNSPECIFIED: IndexOperationMetadataOperationType;
-  CREATING_INDEX: IndexOperationMetadataOperationType;
-  values(): Array<IndexOperationMetadataOperationType>;
-}
-export declare const IndexOperationMetadataOperationTypeEnum: IIndexOperationMetadataOperationTypeEnum;
 export declare type IndexState =
   | 'STATE_UNSPECIFIED'
   | 'CREATING'
@@ -219,6 +212,8 @@ export declare namespace firestoreV1beta1ApiClientInterfaces {
   interface FieldTransform {
     fieldPath?: string;
     setToServerValue?: FieldTransformSetToServerValue;
+    appendMissingElements?: ArrayValue;
+    removeAllFromArray?: ArrayValue;
   }
   interface Filter {
     compositeFilter?: CompositeFilter;
@@ -234,14 +229,6 @@ export declare namespace firestoreV1beta1ApiClientInterfaces {
   interface IndexField {
     fieldPath?: string;
     mode?: IndexFieldMode;
-  }
-  interface IndexOperationMetadata {
-    startTime?: string;
-    endTime?: string;
-    index?: string;
-    operationType?: IndexOperationMetadataOperationType;
-    cancelled?: boolean;
-    documentProgress?: Progress;
   }
   interface LatLng {
     latitude?: number;
@@ -292,10 +279,6 @@ export declare namespace firestoreV1beta1ApiClientInterfaces {
   interface Precondition {
     exists?: boolean;
     updateTime?: string;
-  }
-  interface Progress {
-    workCompleted?: string;
-    workEstimated?: string;
   }
   interface Projection {
     fields?: Array<FieldReference>;
@@ -425,7 +408,6 @@ export declare type FieldTransform = firestoreV1beta1ApiClientInterfaces.FieldTr
 export declare type Filter = firestoreV1beta1ApiClientInterfaces.Filter;
 export declare type Index = firestoreV1beta1ApiClientInterfaces.Index;
 export declare type IndexField = firestoreV1beta1ApiClientInterfaces.IndexField;
-export declare type IndexOperationMetadata = firestoreV1beta1ApiClientInterfaces.IndexOperationMetadata;
 export declare type LatLng = firestoreV1beta1ApiClientInterfaces.LatLng;
 export declare type ListCollectionIdsRequest = firestoreV1beta1ApiClientInterfaces.ListCollectionIdsRequest;
 export declare type ListCollectionIdsResponse = firestoreV1beta1ApiClientInterfaces.ListCollectionIdsResponse;
@@ -437,7 +419,6 @@ export declare type MapValue = firestoreV1beta1ApiClientInterfaces.MapValue;
 export declare type Operation = firestoreV1beta1ApiClientInterfaces.Operation;
 export declare type Order = firestoreV1beta1ApiClientInterfaces.Order;
 export declare type Precondition = firestoreV1beta1ApiClientInterfaces.Precondition;
-export declare type Progress = firestoreV1beta1ApiClientInterfaces.Progress;
 export declare type Projection = firestoreV1beta1ApiClientInterfaces.Projection;
 export declare type QueryTarget = firestoreV1beta1ApiClientInterfaces.QueryTarget;
 export declare type ReadOnly = firestoreV1beta1ApiClientInterfaces.ReadOnly;