+++ packages/messaging/test/controller-interface.test.ts
@@ -18,9 +18,6 @@ import { FirebaseApp } from '@firebase/app-types';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 
-import { makeFakeApp } from './make-fake-app';
-import { makeFakeSWReg } from './make-fake-sw-reg';
-
 import { ControllerInterface } from '../src/controllers/controller-interface';
 import { SWController } from '../src/controllers/sw-controller';
 import { WindowController } from '../src/controllers/window-controller';
@@ -28,6 +25,9 @@ import { DEFAULT_PUBLIC_VAPID_KEY } from '../src/models/fcm-details';
 import { IIDModel } from '../src/models/iid-model';
 import { TokenDetailsModel } from '../src/models/token-details-model';
 import { VapidDetailsModel } from '../src/models/vapid-details-model';
+
+import { makeFakeApp } from './testing-utils/make-fake-app';
+import { makeFakeSWReg } from './testing-utils/make-fake-sw-reg';
 
 const controllersToTest = [WindowController, SWController];
 
@@ -84,8 +84,8 @@ describe('Firebase Messaging > *ControllerInterface', () => {
   });
 
   describe('getPushSubscription()', () => {
-    for (const ControllerInTest of controllersToTest) {
-      it(`should return rejection error in ${ControllerInTest.name}`, () => {
+    for (const controllerInTest of controllersToTest) {
+      it(`should return rejection error in ${controllerInTest.name}`, () => {
         const injectedError = new Error('Inject error.');
         const reg = makeFakeSWReg();
         sandbox.stub(reg, 'pushManager').value({
@@ -94,7 +94,7 @@ describe('Firebase Messaging > *ControllerInterface', () => {
           }
         });
 
-        const controller = new ControllerInTest(app);
+        const controller = new controllerInTest(app);
         return controller
           .getPushSubscription(reg, DEFAULT_PUBLIC_VAPID_KEY)
           .then(
@@ -114,7 +114,7 @@ describe('Firebase Messaging > *ControllerInterface', () => {
           getSubscription: async () => exampleSubscription
         });
 
-        const controller = new ControllerInTest(app);
+        const controller = new controllerInTest(app);
         return controller
           .getPushSubscription(reg, DEFAULT_PUBLIC_VAPID_KEY)
           .then(subscription => {
@@ -127,7 +127,7 @@ describe('Firebase Messaging > *ControllerInterface', () => {
         const reg = makeFakeSWReg();
         sandbox.stub(reg, 'pushManager').value({
           getSubscription: async () => {},
-          subscribe: async options => {
+          subscribe: async (options: PushSubscriptionOptions) => {
             expect(options).to.deep.equal({
               userVisibleOnly: true,
               applicationServerKey: DEFAULT_PUBLIC_VAPID_KEY
@@ -137,7 +137,7 @@ describe('Firebase Messaging > *ControllerInterface', () => {
           }
         });
 
-        const controller = new ControllerInTest(app);
+        const controller = new controllerInTest(app);
         return controller
           .getPushSubscription(reg, DEFAULT_PUBLIC_VAPID_KEY)
           .then(subscription => {