+++ tools/config.js
@@ -63,7 +63,9 @@ Promise.resolve(userToken || cachedToken)
         }
       ]);
 
-      const { projectId: { id } } = response;
+      const {
+        projectId: { id }
+      } = response;
 
       return id;
     })();